network: socket and address build cleanup
- split socket interface from socket
- add default socket interface library
- move io handle to default socket interface library from address
