+++ source/common/network/listen_socket_impl.h
@@ -8,10 +8,10 @@
 #include "envoy/network/connection.h"
 #include "envoy/network/listen_socket.h"
 #include "envoy/network/socket.h"
+#include "envoy/network/socket_interface.h"
 
 #include "common/common/assert.h"
 #include "common/network/socket_impl.h"
-#include "common/network/socket_interface_impl.h"
 
 namespace Envoy {
 namespace Network {
