+++ source/common/network/socket_impl.cc
@@ -1,12 +1,10 @@
 #include "common/network/socket_impl.h"
 
 #include "envoy/common/exception.h"
+#include "envoy/network/socket_interface.h"
 
 #include "common/api/os_sys_calls_impl.h"
 #include "common/common/utility.h"
-#include "common/network/address_impl.h"
-#include "common/network/io_socket_handle_impl.h"
-#include "common/network/socket_interface_impl.h"
 
 namespace Envoy {
 namespace Network {
