+++ source/common/network/socket_interface.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "envoy/config/typed_config.h"
-#include "envoy/network/socket.h"
+#include "envoy/network/socket_interface.h"
 #include "envoy/registry/registry.h"
 #include "envoy/server/bootstrap_extension_config.h"
 
@@ -51,17 +51,6 @@ static inline const SocketInterface* socketInterface(std::string name) {
 
 using SocketInterfaceSingleton = InjectableSingleton<SocketInterface>;
 using SocketInterfaceLoader = ScopedInjectableLoader<SocketInterface>;
-
-/**
- * Create IoHandle for given address
- * @param type type of socket to be requested
- * @param addr address that is gleaned for address type, version and socket interface name
- * @return @ref Network::IoHandlePtr that wraps the underlying socket file descriptor
- */
-static inline IoHandlePtr ioHandleForAddr(Socket::Type type,
-                                          const Address::InstanceConstSharedPtr addr) {
-  return addr->socketInterface().socket(type, addr);
-}
 
 } // namespace Network
 } // namespace Envoy
