+++ source/server/BUILD
@@ -45,6 +45,7 @@ envoy_cc_library(
         "//source/common/common:utility_lib",
         "//source/common/config:runtime_utility_lib",
         "//source/common/config:utility_lib",
+        "//source/common/network:default_socket_interface_lib",
         "//source/common/network:resolver_lib",
         "//source/common/network:socket_interface_lib",
         "//source/common/network:socket_option_factory_lib",
